Fixed bug for binary search and file pointer manipulation.

+++ arXivSearchEngine/Compressor.h
@@ -9,7 +9,7 @@ class Compressor {
 
 public: // build inverted first. build kgram index based on key values of invertedindex AFTER
 
-	static std::list<unsigned char> encodeNumber(uint64_t source);
+	static std::list<unsigned char> encodeNumber(uint32_t source);
 
 	static std::string encode(const std::list<int> &source);
